# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing Kubernetes resources",
        "title": "Phi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kubernetes.Event"
                            }
                        }
                    }
                }
            }
        },
        "/api/resource/{id}/describe": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "summary": "Get describe YAML of a resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/resource/{id}/reconcile": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Start the reconciliation of a resource that supports it",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/resource/{id}/resume": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Resume the reconciliation of a resource that supports it",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/resource/{id}/suspend": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "Suspend the reconciliation of a resource that supports it",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tree": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get tree",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/kubernetes.Resource"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "kubernetes.Condition": {
            "type": "object",
            "properties": {
                "lastTransitionTime": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "kubernetes.DeploymentMetadata": {
            "type": "object",
            "properties": {
                "availableReplicas": {
                    "type": "integer"
                },
                "readyReplicas": {
                    "type": "integer"
                },
                "replicas": {
                    "type": "integer"
                },
                "updatedReplicas": {
                    "type": "integer"
                }
            }
        },
        "kubernetes.Event": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "firstObserved": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "lastObserved": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "resourceUID": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "kubernetes.FluxMetadata": {
            "type": "object",
            "properties": {
                "isReconciling": {
                    "type": "boolean"
                },
                "isSuspended": {
                    "type": "boolean"
                },
                "lastHandledReconcileAt": {
                    "type": "string"
                }
            }
        },
        "kubernetes.GitRepositoryMetadata": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "commit": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "semver": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "kubernetes.HelmReleaseMetadata": {
            "type": "object",
            "properties": {
                "chartName": {
                    "type": "string"
                },
                "chartVersion": {
                    "type": "string"
                },
                "isReconciling": {
                    "type": "boolean"
                },
                "isSuspended": {
                    "type": "boolean"
                },
                "sourceRef": {
                    "$ref": "#/definitions/kubernetes.SourceRef"
                }
            }
        },
        "kubernetes.KustomizationMetadata": {
            "type": "object",
            "properties": {
                "isReconciling": {
                    "type": "boolean"
                },
                "isSuspended": {
                    "type": "boolean"
                },
                "lastAppliedRevision": {
                    "type": "string"
                },
                "lastAttemptedRevision": {
                    "type": "string"
                },
                "lastHandledReconcileAt": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "sourceRef": {
                    "$ref": "#/definitions/kubernetes.SourceRef"
                }
            }
        },
        "kubernetes.OCIRepositoryMetadata": {
            "type": "object",
            "properties": {
                "digest": {
                    "type": "string"
                },
                "semver": {
                    "type": "string"
                },
                "semverFilter": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "kubernetes.PVCMetadata": {
            "type": "object",
            "properties": {
                "accessModes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "capacity": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "phase": {
                    "type": "string"
                },
                "storageClass": {
                    "type": "string"
                },
                "volumeMode": {
                    "type": "string"
                },
                "volumeName": {
                    "type": "string"
                }
            }
        },
        "kubernetes.PodMetadata": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                }
            }
        },
        "kubernetes.Resource": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kubernetes.Resource"
                    }
                },
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kubernetes.Condition"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deploymentMetadata": {
                    "$ref": "#/definitions/kubernetes.DeploymentMetadata"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kubernetes.Event"
                    }
                },
                "fluxMetadata": {
                    "$ref": "#/definitions/kubernetes.FluxMetadata"
                },
                "gitRepositoryMetadata": {
                    "$ref": "#/definitions/kubernetes.GitRepositoryMetadata"
                },
                "group": {
                    "type": "string"
                },
                "helmReleaseMetadata": {
                    "$ref": "#/definitions/kubernetes.HelmReleaseMetadata"
                },
                "isFluxManaged": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "kustomizationMetadata": {
                    "$ref": "#/definitions/kubernetes.KustomizationMetadata"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ociRepositoryMetadata": {
                    "$ref": "#/definitions/kubernetes.OCIRepositoryMetadata"
                },
                "parentIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "parentRefs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "podMetadata": {
                    "$ref": "#/definitions/kubernetes.PodMetadata"
                },
                "pvcMetadata": {
                    "$ref": "#/definitions/kubernetes.PVCMetadata"
                },
                "resource": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/kubernetes.Status"
                },
                "uid": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "kubernetes.SourceRef": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "kubernetes.Status": {
            "type": "string",
            "enum": [
                "unknown",
                "success",
                "failed",
                "pending",
                "warning"
            ],
            "x-enum-varnames": [
                "StatusUnknown",
                "StatusSuccess",
                "StatusFailed",
                "StatusPending",
                "StatusWarning"
            ]
        }
    }
}